+++ bindings/python/pymongoarrow/types.py
@@ -15,7 +15,7 @@ import enum
 from datetime import datetime
 
 import pyarrow.types as _atypes
-from bson import Int64, ObjectId
+from bson import Decimal128, Int64, ObjectId
 from pyarrow import DataType as _ArrowDataType
 from pyarrow import PyExtensionType, binary, bool_, float64, int64, string, timestamp
 
@@ -28,6 +28,7 @@ class _BsonArrowTypes(enum.Enum):
     objectid = 5
     string = 6
     bool = 7
+    decimal128_str = 8
 
 
 # Custom Extension Types.
@@ -45,6 +46,17 @@ class ObjectIdType(PyExtensionType):
         return ObjectIdType, ()
 
 
+class Decimal128StringType(PyExtensionType):
+    _type_marker = _BsonArrowTypes.decimal128_str
+
+    def __init__(self):
+        # Delayed to prevent circular import.
+        super().__init__(string())
+
+    def __reduce__(self):
+        return Decimal128StringType, ()
+
+
 # Internal Type Handling.
 
 
@@ -53,14 +65,20 @@ def _is_objectid(obj):
     return type_marker == ObjectIdType._type_marker
 
 
+def _is_decimal128_str(obj):
+    type_marker = getattr(obj, "_type_marker", "")
+    return type_marker == Decimal128StringType._type_marker
+
+
 _TYPE_NORMALIZER_FACTORY = {
     Int64: lambda _: int64(),
     float: lambda _: float64(),
     int: lambda _: int64(),
     datetime: lambda _: timestamp("ms"),  # TODO: add tzinfo support
     ObjectId: lambda _: ObjectIdType(),
-    str: lambda: string(),
-    bool: lambda: bool_(),
+    Decimal128: lambda _: Decimal128StringType(),
+    str: lambda _: string(),
+    bool: lambda _: bool_(),
 }
 
 
@@ -70,6 +88,7 @@ _TYPE_CHECKER_TO_INTERNAL_TYPE = {
     _atypes.is_float64: _BsonArrowTypes.double,
     _atypes.is_timestamp: _BsonArrowTypes.datetime,
     _is_objectid: _BsonArrowTypes.objectid,
+    _is_decimal128_str: _BsonArrowTypes.decimal128_str,
     _atypes.is_string: _BsonArrowTypes.string,
     _atypes.is_boolean: _BsonArrowTypes.bool,
 }